BR-53 Write a documentation and README
