+++ src/main/java/svydovets/core/context/AnnotationConfigApplicationContext.java
@@ -2,7 +2,7 @@ package svydovets.core.context;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import svydovets.core.context.beanFactory.BeanFactory;
+import svydovets.core.context.beanFactory.BeanFactoryImpl;
 
 import java.util.Map;
 
@@ -10,35 +10,35 @@ public class AnnotationConfigApplicationContext implements ApplicationContext {
 
     private static final Logger log = LoggerFactory.getLogger(AnnotationConfigApplicationContext.class);
 
-    protected final BeanFactory beanFactory = new BeanFactory();
+    protected final BeanFactoryImpl beanFactoryImpl = new BeanFactoryImpl();
 
     public AnnotationConfigApplicationContext(String basePackage) {
         log.info("Start creating an application context");
-        beanFactory.registerBeans(basePackage);
+        beanFactoryImpl.registerBeans(basePackage);
         log.info("Finish creating an application context");
     }
 
     public AnnotationConfigApplicationContext(Class<?>... classes) { // @Config or @Component
-        beanFactory.registerBeans(classes);
+        beanFactoryImpl.registerBeans(classes);
     }
 
     @Override
     public <T> T getBean(Class<T> requiredType) {
-        return beanFactory.getBean(requiredType);
+        return beanFactoryImpl.getBean(requiredType);
     }
 
     @Override
     public <T> T getBean(String name, Class<T> requiredType) {
-        return beanFactory.getBean(name, requiredType);
+        return beanFactoryImpl.getBean(name, requiredType);
     }
 
     @Override
     public <T> Map<String, T> getBeansOfType(Class<T> requiredType) {
-        return beanFactory.getBeansOfType(requiredType);
+        return beanFactoryImpl.getBeansOfType(requiredType);
     }
 
     @Override
     public Map<String, Object> getBeans() {
-        return beanFactory.getBeans();
+        return beanFactoryImpl.getBeans();
     }
 }
